requirements.txt
openai>=1.0.0
pillow
opencv-python
numpy

